requirements-prod.txt
# Production Requirements for Django E-commerce Store
# Include base requirements
-r requirements.txt

# Production Server
gunicorn==21.2.0

# Static Files
whitenoise==6.6.0

# Database (Choose one based on your needs)
# PostgreSQL (Recommended for production)
psycopg2-binary==2.9.9

# MySQL (Alternative)
# mysqlclient==2.2.0

# Caching
redis==5.0.1

# Background Tasks
celery==5.3.4

# Cloud Storage (AWS S3)
django-storages==1.14.2
boto3==1.34.0

# Email
django-anymail==10.1

# Security
django-cors-headers==4.3.1
django-ratelimit==4.1.0

# Monitoring
sentry-sdk==1.38.0

# Environment
python-decouple==3.8



